+++ lino/core/requests.py
@@ -43,6 +43,7 @@ from lino.core.dbutils import resolve_model, resolve_app
 from lino.core import layouts
 #~ from lino.core import changes
 from lino.core import fields
+from lino.core import actions
 
 #~ from lino.core.perms import UserLevels
 #~ from lino.core import perms 
@@ -90,6 +91,12 @@ class EmptyTableRow(VirtualRow):
     
     def filename_root(self):
         return self._table.app_label + '.' + self._table.__name__
+        
+    def __getattr__(self,name):
+        v = getattr(self._table,name)
+        if isinstance(v,actions.Action):
+            return actions.InstanceAction(v,self._table,self,None)
+        return v